LIBASPACE-146 Update Vagrant box
This updates the aspace vagrant box is use the last version in the
puppetlabs repository. Also includes a line to update the packages (
except the puppet one because the repo is out of date ). This should fix
the issue with NSS not being able to process the github SSL cert.

https://issues.umd.edu/browse/LIBASPACE-146