LIBASPACE-63 Use Docker configuration
This sets up the Vagrant to use the Dockerized Aspace ENV.
The configuration can be see in the docker-compose-vagrant.yml.
The vagrant uses the Aspace, Nginx, and MySQL containers, but not the
Solr container. Solr is still setup in its own Vagrant VM, which aspace
is configured to use in the docker-compose.yml file.

https://issues.umd.edu/browse/LIBASPACE-163